feat: Find the closest node
Find the closest point to user input node. Many nodes in the OSM data are on
closed paths or may be isolated is some other way. If user choose one one of
these nodes, the program may not be able to find a path from that node to the
end node. To avoid this, the program will choose the closest node that is on a
road that is not a footway.